Moved more logic to service
