+++ src/app/app-config.model.ts
@@ -1,7 +1,6 @@
 export interface AppConfig {
     apiServerGqlUrl: string;
     apiServerHttpUrl: string;
-    singleTenant: boolean;
     githubClientId?: string;
     featureFlags: AppConfigFeatureFlags;
     loginInstructions?: AppConfigLoginInstructions;
