+++ src/app/app-config.service.ts
@@ -33,14 +33,6 @@ export class AppConfigService {
         return this.appConfig.apiServerHttpUrl;
     }
 
-    get singleTenant(): boolean {
-        if (!this.appConfig) {
-            this.appConfig = AppConfigService.loadAppConfig();
-        }
-
-        return this.appConfig.singleTenant;
-    }
-
     get githubClientId(): MaybeUndefined<string> {
         if (!this.appConfig) {
             this.appConfig = AppConfigService.loadAppConfig();
