+++ src/app/dataset-view/additional-components/overview-component/components/add-data-modal/add-data-modal.component.spec.ts
@@ -1,23 +1,29 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { SharedTestModule } from "src/app/common/shared-test.module";
+import { Apollo } from "apollo-angular";
+import { ComponentFixture, TestBed } from "@angular/core/testing";
+import { AddDataModalComponent } from "./add-data-modal.component";
+import { NgbActiveModal } from "@ng-bootstrap/ng-bootstrap";
+import { HttpClientModule } from "@angular/common/http";
+import { MatIconModule } from "@angular/material/icon";
+import { MatDividerModule } from "@angular/material/divider";
 
-import { AddDataModalComponent } from './add-data-modal.component';
+describe("AddDataModalComponent", () => {
+    let component: AddDataModalComponent;
+    let fixture: ComponentFixture<AddDataModalComponent>;
 
-describe('AddDataModalComponent', () => {
-  let component: AddDataModalComponent;
-  let fixture: ComponentFixture<AddDataModalComponent>;
+    beforeEach(async () => {
+        await TestBed.configureTestingModule({
+            declarations: [AddDataModalComponent],
+            providers: [Apollo, NgbActiveModal],
+            imports: [HttpClientModule, SharedTestModule, MatIconModule, MatDividerModule],
+        }).compileComponents();
 
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ AddDataModalComponent ]
-    })
-    .compileComponents();
+        fixture = TestBed.createComponent(AddDataModalComponent);
+        component = fixture.componentInstance;
+        fixture.detectChanges();
+    });
 
-    fixture = TestBed.createComponent(AddDataModalComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
+    it("should create", () => {
+        expect(component).toBeTruthy();
+    });
 });
