+++ src/app/dataset-view/additional-components/overview-component/components/file-from-url-modal/file-from-url-modal.component.spec.ts
@@ -1,23 +1,27 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from "@angular/core/testing";
+import { FileFromUrlModalComponent } from "./file-from-url-modal.component";
+import { NgbActiveModal } from "@ng-bootstrap/ng-bootstrap";
+import { ReactiveFormsModule } from "@angular/forms";
+import { SharedTestModule } from "src/app/common/shared-test.module";
+import { MatDividerModule } from "@angular/material/divider";
 
-import { FileFromUrlModalComponent } from './file-from-url-modal.component';
+describe("FileFromUrlModalComponent", () => {
+    let component: FileFromUrlModalComponent;
+    let fixture: ComponentFixture<FileFromUrlModalComponent>;
 
-describe('FileFromUrlModalComponent', () => {
-  let component: FileFromUrlModalComponent;
-  let fixture: ComponentFixture<FileFromUrlModalComponent>;
+    beforeEach(async () => {
+        await TestBed.configureTestingModule({
+            declarations: [FileFromUrlModalComponent],
+            providers: [NgbActiveModal],
+            imports: [ReactiveFormsModule, SharedTestModule, MatDividerModule],
+        }).compileComponents();
 
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ FileFromUrlModalComponent ]
-    })
-    .compileComponents();
+        fixture = TestBed.createComponent(FileFromUrlModalComponent);
+        component = fixture.componentInstance;
+        fixture.detectChanges();
+    });
 
-    fixture = TestBed.createComponent(FileFromUrlModalComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
+    it("should create", () => {
+        expect(component).toBeTruthy();
+    });
 });
