+++ src/app/services/file-upload.service.ts
@@ -1,10 +1,16 @@
 import { AppConfigService } from "src/app/app-config.service";
 import { HttpClient, HttpHeaders } from "@angular/common/http";
-import { Injectable } from "@angular/core";
-import { Observable } from "rxjs";
-import { MaybeUndefined, UploadPrepareResponse } from "../common/app.types";
+import { Injectable, Injector } from "@angular/core";
+import { Observable, finalize, of, switchMap, tap } from "rxjs";
+import { MaybeUndefined, UploadAvailableMethod, UploadPerareData, UploadPrepareResponse } from "../common/app.types";
 import { LocalStorageService } from "./local-storage.service";
 import { DatasetInfo } from "../interface/navigation.interface";
+import { DatasetBasicsFragment } from "../api/kamu.graphql.interface";
+import { APOLLO_OPTIONS } from "apollo-angular";
+import { DatasetApi } from "../api/dataset.api";
+import { DatasetViewTypeEnum } from "../dataset-view/dataset-view.interface";
+import { DatasetService } from "../dataset-view/dataset.service";
+import { NavigationService } from "./navigation.service";
 
 @Injectable({
     providedIn: "root",
@@ -13,15 +19,79 @@ export class FileUploadService {
     private authHeaders: MaybeUndefined<{
         [header: string]: string | string[];
     }>;
+    private uploadToken: string;
+
     constructor(
         private http: HttpClient,
         private localStorageService: LocalStorageService,
         private appConfigService: AppConfigService,
+        private navigationService: NavigationService,
+        private datasetService: DatasetService,
+        private injector: Injector,
     ) {
         this.authHeaders = { Authorization: `Bearer ${this.localStorageService.accessToken}` };
     }
 
-    public uploadFilePrepare(file: File): Observable<UploadPrepareResponse> {
+    public uploadFile(file: File, datasetBasics: DatasetBasicsFragment): Observable<object> {
+        const uploadPrepare$: Observable<UploadPrepareResponse> = this.uploadFilePrepare(file);
+        return uploadPrepare$.pipe(
+            tap((data) => (this.uploadToken = data.uploadToken)),
+            switchMap((uploadPrepareResponse: UploadPrepareResponse) =>
+                this.prepareUploadData(uploadPrepareResponse, file),
+            ),
+            switchMap(({ uploadPrepareResponse, bodyObject, uploadHeaders }: UploadPerareData) =>
+                this.uploadFileByMethod(
+                    uploadPrepareResponse.method,
+                    uploadPrepareResponse.uploadUrl,
+                    bodyObject,
+                    uploadHeaders,
+                ),
+            ),
+            switchMap(() =>
+                this.ingestDataToDataset(
+                    {
+                        accountName: datasetBasics.owner.accountName,
+                        datasetName: datasetBasics.name,
+                    },
+                    this.uploadToken,
+                ),
+            ),
+            finalize(() => {
+                this.updatePage(datasetBasics);
+            }),
+        );
+    }
+
+    private updatePage(datasetBasics: DatasetBasicsFragment): void {
+        this.updateCache(datasetBasics);
+        this.datasetService
+            .requestDatasetMainData({
+                accountName: datasetBasics.owner.accountName,
+                datasetName: datasetBasics.name,
+            })
+            .subscribe(),
+            this.navigationService.navigateToDatasetView({
+                accountName: datasetBasics.owner.accountName,
+                datasetName: datasetBasics.name,
+                tab: DatasetViewTypeEnum.Overview,
+            });
+    }
+
+    private updateCache(datasetBasics: DatasetBasicsFragment): void {
+        const cache = this.injector.get(APOLLO_OPTIONS).cache;
+        if (cache) {
+            const datasetKeyFragment = DatasetApi.generateDatasetKeyFragment(
+                cache.identify(DatasetApi.generateAccountKeyFragment(datasetBasics.owner.id)),
+                datasetBasics.id,
+            );
+            cache.evict({
+                id: cache.identify(datasetKeyFragment),
+                fieldName: "data",
+            });
+        }
+    }
+
+    private uploadFilePrepare(file: File): Observable<UploadPrepareResponse> {
         return this.http.post<UploadPrepareResponse>(
             `${this.appConfigService.apiServerHttpUrl}/platform/file/upload/prepare?fileName=` +
                 file.name +
@@ -34,20 +104,20 @@ export class FileUploadService {
         );
     }
 
-    public uploadPostFile(url: string, bodyObject: File | FormData, uploadHeaders: HttpHeaders): Observable<object> {
+    private uploadPostFile(url: string, bodyObject: File | FormData, uploadHeaders: HttpHeaders): Observable<object> {
         return this.http.post(url, bodyObject, {
             headers: uploadHeaders,
         });
     }
 
-    public uploadPutFile(url: string, bodyObject: File | FormData, uploadHeaders: HttpHeaders): Observable<object> {
+    private uploadPutFile(url: string, bodyObject: File | FormData, uploadHeaders: HttpHeaders): Observable<object> {
         return this.http.put(url, bodyObject, {
             headers: uploadHeaders,
         });
     }
 
-    public ingestDataToDataset(datasetInfo: DatasetInfo, uploadToken: string): Observable<object> {
-        const datasetAlias = this.appConfigService.singleTenant
+    private ingestDataToDataset(datasetInfo: DatasetInfo, uploadToken: string): Observable<object> {
+        const datasetAlias = this.appConfigService.featureFlags.enableLogout
             ? datasetInfo.datasetName
             : `${datasetInfo.accountName}/${datasetInfo.datasetName}`;
         return this.http.post<object>(
@@ -55,5 +125,44 @@ export class FileUploadService {
             null,
             { headers: this.authHeaders },
         );
+    }
+
+    private prepareUploadData(uploadPrepareResponse: UploadPrepareResponse, file: File): Observable<UploadPerareData> {
+        let uploadHeaders = new HttpHeaders();
+        uploadPrepareResponse.headers.forEach((header: [string, string]) => {
+            uploadHeaders = uploadHeaders.append(header[0], header[1]);
+        });
+        let bodyObject: FormData | File;
+        if (uploadPrepareResponse.useMultipart) {
+            const formData = new FormData();
+            uploadPrepareResponse.fields.forEach((field: [string, string]) => {
+                formData.append(field[0], field[1]);
+            });
+            formData.append("file", file);
+            bodyObject = formData;
+        } else {
+            bodyObject = file;
+        }
+        return of({
+            uploadPrepareResponse,
+            bodyObject,
+            uploadHeaders,
+        });
+    }
+
+    private uploadFileByMethod(
+        method: UploadAvailableMethod,
+        uploadUrl: string,
+        bodyObject: File | FormData,
+        uploadHeaders: HttpHeaders,
+    ): Observable<object> {
+        switch (method) {
+            case "POST":
+                return this.uploadPostFile(uploadUrl, bodyObject, uploadHeaders);
+            case "PUT":
+                return this.uploadPutFile(uploadUrl, bodyObject, uploadHeaders);
+            default:
+                throw new Error("Unexpected upload method");
+        }
     }
 }
